requirements.txt
playwright==1.42.0
python-telegram-bot==21.0.1
openai==1.14.0
python-dotenv==1.0.1
